requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
defusedxml==0.7.1
fonttools==4.60.1
fpdf2==2.8.5
idna==3.11
lxml==6.0.2
pillow==12.0.0
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6_sip==13.10.2
PyYAML==6.0.3
requests==2.32.5
urllib3==2.5.0
